+++ includes/crawler.h
@@ -1,0 +1,10 @@
+#ifndef CRAWLER_H
+#define CRAWLER_H
+
+#include <vector>
+#include <string>
+#include <iostream>
+
+std::vector<std::string> crawl_files();
+
+#endif // CRAWLER_H
