+++ includes/reader.h
@@ -1,0 +1,10 @@
+#ifndef READER_H
+#define READER_H
+
+#include <vector>
+#include <string>
+#include <regex>
+
+std::vector<std::string> process_file(const std::string& line);
+
+#endif // READER_H
